+++ packages/observer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rabjs/observer
 
+## 9.0.1
+
+### Patch Changes
+
+- fix: $model access
+
 ## 9.0.0
 
 ### Major Changes