+++ packages/service/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rabjs/service
 
+## 9.0.1
+
+### Patch Changes
+
+- fix: $model access
+
 ## 9.0.0
 
 ### Major Changes